+++ NEWS.md
@@ -1,6 +1,7 @@
 #Version 0.1.6 [development]
 ### Minor changes
 * Fixes issue #22 - wLambda error when a class has less than 2 instances
+* Fixes infotheo when the dataset cannot be discretized
 
 
 # Version 0.1.5 [current]